ordofp::gat

Trait Apply

pub trait Apply: Functor {
    // Required method
    fn apply<B, F>(self, f: Self::Target<F>) -> Self::Target<B>
       where F: FnMut(Self::Inner) -> B;
}

Apply extends Functor with the ability to apply a wrapped function to a wrapped value.

Required Methods

fn apply<B, F>(self, f: Self::Target<F>) -> Self::Target<B>
where F: FnMut(Self::Inner) -> B,

Applies a wrapped function to this value.

The function is wrapped in the same context as the value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.